refactor: extract dependency graph helper (#236)
