+++ packages/dependencies/package.json
@@ -20,6 +20,7 @@
     "prepack": "run workspace:build \"$(pwd)\""
   },
   "dependencies": {
+    "@yarnpkg/core": "^2.4.0",
     "monodeploy-types": "workspace:packages/types"
   }
 }
