+++ packages/node/package.json
@@ -35,6 +35,7 @@
     "conventional-changelog-writer": "^5.0.0",
     "conventional-commits-parser": "^3.2.0",
     "monodeploy-changelog": "workspace:packages/changelog",
+    "monodeploy-dependencies": "workspace:packages/dependencies",
     "monodeploy-git": "workspace:packages/git",
     "monodeploy-io": "workspace:packages/io",
     "monodeploy-logging": "workspace:packages/logging",
